+++ Infrastructure/Immobilier.Infrastructure/Repository/Contracts/IUserRepository.cs
@@ -1,16 +1,14 @@
 ﻿using Immobilier.Domain;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace Immobilier.Infrastructure.Repository.Contracts
 {
     public interface IUserRepository
     {
-        ulong CreateUser(string name, string email, string password, int age);
-        Task<User?> GetUserById(ulong userId);
+        uint CreateUser(string name, string email, string password, int age);
+        Task<User?> GetUserById(uint userId);
         Task<User?> GetAuthenticatedUser(string email, string password);
         Task<User?> GetUserByEmail(string email);
         Task<IEnumerable<User>> GetAllUsers();
-        Task<User?> UpdateUser(ulong id, string name, string email, uint age);
+        Task<User?> UpdateUser(uint id, string name, string email, uint age);
     }
 }
