+++ Infrastructure/Immobilier.Infrastructure/Repository/UserRepository.cs
@@ -2,8 +2,6 @@
 using Immobilier.Infrastructure.Config;
 using Immobilier.Infrastructure.Repository.Contracts;
 using Microsoft.EntityFrameworkCore;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace Immobilier.Infrastructure.Repository
 {
@@ -16,7 +14,7 @@ namespace Immobilier.Infrastructure.Repository
             _context = context;
         }
 
-        public ulong CreateUser(string name, string email, string password, int age)
+        public uint CreateUser(string name, string email, string password, int age)
         {
             var userToCreate = new User(name, password, email, age);
             var newUser = _context.Users.Add(userToCreate).Entity;
@@ -31,7 +29,7 @@ namespace Immobilier.Infrastructure.Repository
             return await _context.Users.AsNoTracking().ToArrayAsync();
         }
 
-        public async Task<User?> GetUserById(ulong userId)
+        public async Task<User?> GetUserById(uint userId)
         {
             return await _context.Users.AsNoTracking().SingleOrDefaultAsync(u => u.Id == userId);
         }
@@ -46,7 +44,7 @@ namespace Immobilier.Infrastructure.Repository
             return await _context.Users.AsNoTracking().SingleOrDefaultAsync(u => u.Email == email);
         }
 
-        public async Task<User?> UpdateUser(ulong id, string name, string email, uint age)
+        public async Task<User?> UpdateUser(uint id, string name, string email, uint age)
         {
             var userToEdit = await _context.Users.FirstOrDefaultAsync(u => u.Id == id);
             if (userToEdit == null) return null;
